[PN-90] Add nodelets to create PointCloud2 messages from depth_registered and rgb image topics (#147)
* Replaces the previous use_depth_registered_nodelets boolean launch
argument with the new depth_registered_mode string launch argument.
This is converted into an enum within the spot_driver.launch.py launch
file. This simplifies handling the three possible configurations of
registered depth image publishing, and makes the configuration of
these options more obvious to the user.

* Adds a new publish_point_clouds launch argument. When this argument
is set to true, the driver launch file will add new nodelets that
subscribe to the depth_registered and RGB image topics for each Spot
camera and use these images to create and publish point clouds.

* Changes the executor type of the composable node container to
component_container_mt. This is needed to allow the increased number
of nodelets to work in parallel without blocking each other.

* Moves creation of the image processing nodelets to within
spot_driver.launch.py instead of a separate launch file. This
facilitates reuse of the same composable node container for different
collections of nodelets.