[SW-679] Config. file is not used to get hostname, username, or password when running the driver (#259)
## Changes Overview
This addresses [this issue](https://github.com/bdaiinstitute/spot_ros2/issues/247) of not being able to connect to Spot from `spot_driver.launch.py` without setting environment variables. Now, the spot driver can be launched with only a configuration file yaml (`config_file`) if the user chooses. 

The implementation parses the `config_file` yaml within the launchfile to determine if `username`,  `password`, and `hostname` are set there as ROS parameters. If both environment variables and `config_file` parameters are set, then the environment variables take priority (similar to the `get_from_env_and_fall_back_to_param` logic in `spot_ros2.py`).  It also assumes a yaml structure for `config_file` (if used) of the following:
```
/**:
    ros__parameters:
        username: ...
        password: ...
        hostname: xxx.xxx.xxx.xxx
        .....
```
and also assumes environment variables (if used) are set as `BOSDYN_CLIENT_USERNAME`, `BOSDYN_CLIENT_PASSWORD,` and `SPOT_IP`. 


## Testing Done
* If an environment variable (any of `BOSDYN_CLIENT_USERNAME`, `BOSDYN_CLIENT_PASSWORD,` `SPOT_IP`) is set, this will be used for the login attempt. 
* If any environment variable is not set, the `config_file` yaml will be searched to see if the appropriate login information exists there. 
* If a parameter of username/password/hostname is set both by environment variable and by `config_file`, the value from the environment variable will be used for the login attempt. 
* If any one of username, password, and hostname is not set by anything, an exception will be thrown as login will not be possible and the driver will not launch. 
* If the `config_file` yaml passed in is malformed and cannot be parsed, a YAMLError is thrown. 

Verified that login is possible with a config file on-robot too. 

## Other comments
Currently the only use for these variables in the launchfile is to instantiate SpotWrapper in order to test if the robot has an arm. This affects which nodes get launched (specifically the robot description and the camera/point cloud nodes to stream from the hand camera). I feel that this implementation results in a messy launchfile, and also means the robot is logged into twice (once in the launchfile, and then again in the main `spot_ros2` node). Parsing a configuration file by hand is the only way to get the value of these parameters in the launchfile, as the standard ROS way of getting parameters from a yaml is only available for nodes. It would be nice in the future to try to rework this to make for a more declarative launchfile, potentially using things like ActionHandlers.