[SW-817] Unify login parameter names for launching driver from a config file (#292)
## Change Overview

When launching the spot driver, there is logic to get the username, password, and spot IP address from an environment variable, and if they are not found, fall back to a ROS parameter. The `spot_ros2` node assumes that Spot's IP address, if specified in a config file, is listed as `hostname` whereas the `rclcpp_parameter_interface` in `spot_driver` assumes it is listed as `address`. This changes all instances of `address` to `hostname` to make launching with a config file fully functional and match the config file template given in [`spot_driver/config/spot_ros_example.yaml`](https://github.com/bdaiinstitute/spot_ros2/blob/main/spot_driver/config/spot_ros_example.yaml). 

## Testing Done

- [x] Successfully able to launch the driver and log in with the hostname set in a configuration file, and not as an environment variable. From here, was able to run the examples `walk_forward` and `send_inverse_kinematics_examples`. 
- [x] tests passed with updated function names