[SW-858] Remove python based image publishers (#305)
## Change Overview

The python based image publishers currently don't launch and have been replaced for a while with the faster C++ image publishers. This PR deletes these Python image publishers and updates the README of spot driver with some more information of how the driver works. 

Additionally, this PR deletes an empty launchfile `view_robot.launch.py` and deletes the `.vscode` folder (which shouldn't have been there in the first place as it's in the gitignore). 

## Testing Done

- [x] CI
- [x] already addressed possible downstream effects of deleting these launchfiles in our internal repos in previous PRs.