[SW-986] remove command_spot node and update README with more documentation (#357)
## Change Overview

This PR deletes the `command_spot` node as it currently doesn't work, and hasn't been maintained in a while (see https://github.com/bdaiinstitute/spot_ros2/issues/351).

Also, this cleans up the section on the driver in the main README and points people to using service calls from the driver, as this was the intended use case of the `command_spot` node. Added some more info in the help section with common issues too. 

## Testing Done
- [x] `command_spot` node is no longer an option to run after building.