[PN-60] Add node to stitch front camera images (#332)
## Change Overview

Adds `image_stitcher_node` to `spot_driver` to stitch together the front camera images and publish the stitched image and the transform to the virtual camera.

This node can get launched now when running `spot_driver.launch.py` with the launch argument `stitch_front_images:=True` (it is disabled by default). Default parameters for the image stitcher node have also been added to the example `spot_ros_example.yaml` configuration file.   Users will need to update their configuration files with these values in order to receive any meaningful stitched images as we do not launch the driver with a default configuration file. 

## Testing Done
On robot testing was successfully able to view the stitched image:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Xynj6CBpA3NqqBFfE8Q9/d07a8a81-4c34-4f30-ab5a-4296494abedb.png)

The stitched image comes in at around 1/2 the speed as the standard camera streams on my system. If this is an issue for users we can address it in a future PR. 

The default parameters for the virtual camera were verified as reasonable on 3 different robots (Opal, Lionel, and Pluto).